+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Columns can be aligned by a column spec, e.g. `#{key => foo, alight => right}`
+
 ## [0.1.0] - 2022-02-28
 
 Initial version.